+++ core/src/main/java/org/web3j/tx/RawTransactionManager.java
@@ -127,7 +127,7 @@ public class RawTransactionManager extends TransactionManager {
 
     @Override
     public EthSendTransaction sendEIP1559Transaction(
-            BigInteger chainId,
+            long chainId,
             BigInteger maxPriorityFeePerGas,
             BigInteger maxFeePerGas,
             BigInteger gasLimit,
